Refactor analog control dialog into package with reusable widgets
- Extract AnalogControlDialog to analog_control/ package
- Add position_capture module for screen coordinate capture
- Add spin_inputs module for compact integer entry and spin rows
- Split dialog groups and options into separate modules
