+++ keymasq/gui/widgets/analog_control/__init__.py
@@ -1,0 +1,3 @@
+from keymasq.gui.widgets.analog_control.dialog import AnalogControlDialog
+
+__all__ = ["AnalogControlDialog"]
