+++ keymasq/gui/widgets/gamepad_output_choices.py
@@ -1,5 +1,6 @@
 import logging
 from collections.abc import Callable, Sequence
+from dataclasses import dataclass
 from typing import Protocol
 
 from keymasq.common.devices import is_gamepad_button_name
@@ -13,8 +14,21 @@ from keymasq.session.settings import load_virtual_gamepad_count
 log = logging.getLogger("keymasq.gui.widgets.gamepad_output_choices")
 
 
-class _HardwareManagerLike(Protocol):
+@dataclass(frozen=True, slots=True)
+class GamepadOutputChoiceSet:
+    choices: list[tuple[str | None, str]]
+    count: int
+    hardware_configs: list[object]
+
+
+class HardwareManagerLike(Protocol):
     def list_hardware(self) -> Sequence[object]: ...
+
+
+class WarningLabelLike(Protocol):
+    def set_label(self, label: str) -> None: ...
+
+    def set_visible(self, visible: bool) -> None: ...
 
 
 def virtual_gamepad_count() -> int:
@@ -114,11 +128,10 @@ def gamepad_output_choices_for(
 
 
 def load_gamepad_output_hardware_configs(
-    hardware_manager_factory: Callable[[], _HardwareManagerLike] = HardwareManager,
+    hardware_manager_factory: Callable[[], HardwareManagerLike] = HardwareManager,
 ) -> list[object]:
     try:
-        manager = hardware_manager_factory()
-        return list(manager.list_hardware())
+        return list(hardware_manager_factory().list_hardware())
     except (OSError, RuntimeError) as exc:
         log.debug("Unable to load hardware configs for gamepad outputs: %s", exc)
         return []
@@ -129,10 +142,48 @@ def gamepad_output_choices(
     *,
     count: int | None = None,
     hardware_configs: Sequence[object] | None = None,
-    hardware_manager_factory: Callable[[], _HardwareManagerLike] = HardwareManager,
+    hardware_manager_factory: Callable[[], HardwareManagerLike] = HardwareManager,
 ) -> list[tuple[str | None, str]]:
     if count is None:
         count = virtual_gamepad_count()
     if hardware_configs is None:
         hardware_configs = load_gamepad_output_hardware_configs(hardware_manager_factory)
     return gamepad_output_choices_for(selected_id, count, hardware_configs)
+
+
+def load_gamepad_output_choices(
+    selected_id: str | None,
+    *,
+    count_loader: Callable[[], int] = virtual_gamepad_count,
+    hardware_manager_factory: Callable[[], HardwareManagerLike] = HardwareManager,
+) -> GamepadOutputChoiceSet:
+    count = count_loader()
+    hardware_configs = load_gamepad_output_hardware_configs(hardware_manager_factory)
+    return GamepadOutputChoiceSet(
+        choices=gamepad_output_choices_for(selected_id, count, hardware_configs),
+        count=count,
+        hardware_configs=hardware_configs,
+    )
+
+
+def selected_gamepad_output_id(
+    dropdown_selected: int,
+    output_ids: Sequence[str | None],
+    current_output_id: str | None,
+) -> str | None:
+    if 0 <= dropdown_selected < len(output_ids):
+        return output_ids[dropdown_selected]
+    return current_output_id
+
+
+def update_gamepad_output_warning_label(
+    label: WarningLabelLike | None,
+    output_id: str | None,
+    *,
+    count_loader: Callable[[], int] = virtual_gamepad_count,
+) -> str | None:
+    message = gamepad_output_unavailable_message(output_id, count_loader())
+    if label is not None:
+        label.set_label(message or "")
+        label.set_visible(bool(message))
+    return message
